media/requirements.txt
paramiko==3.2.0
python-dotenv==1.0.0
Requests==2.31.0
